New stories file for DailyForecastDisplay; move common storybook functions to /utils
